Add tech stack mapping and refactor Project and WorkItem pages to utilize it
