+++ src/data/work/SmallBossCorp.mdx
@@ -1,17 +1,35 @@
 ---
 name: Small Boss Corp
-description: |
-  Small Boss Corp is a small company that specializes in creating and selling high-quality, affordable products. They are known for their innovative designs and commitment to customer satisfaction. The company has a strong online presence and offers a wide range of products, including electronics, home goods, and fashion accessories.
-  Small Boss Corp is dedicated to sustainability and ethical business practices. They source materials responsibly and prioritize eco-friendly packaging. The company also supports local communities by partnering with artisans and small businesses.
-  With a focus on quality and affordability, Small Boss Corp aims to make a positive impact on the world while providing customers with the best products possible.
+description: Small Boss Corp is a small company that specializes in creating and selling high-quality, affordable products. They are known for their innovative designs and commitment to customer satisfaction.
 techStack:
-    - React
-    - TypeScript
-    - TailwindCSS
-    - Node.js
-    - PostgreSQL
-    - Azure
-position: Full-Stack Developer
+  - React
+  - TypeScript
+  - TailwindCSS
+  - Node.js
+  - PostgreSQL
+  - Azure
 ---
 
 # About Small Boss Corp
+
+Small Boss Corp is dedicated to sustainability and ethical business practices...
+
+## Roles
+
+### Junior Software Engineer
+
+**Duration:** 11/2022 - 05/2023
+**Location:** San Francisco, CA
+
+- Developed and maintained web applications using **React** and **TypeScript**.
+- Collaborated with cross-functional teams to design and implement new features.
+
+### Intern
+
+**Duration:** 06/2022 - 10/2022
+**Location:** Remote
+
+- Helped in the development of web applications.
+- Participated in code reviews.
+- Gained experience in Agile methodologies and version control systems.
+
