+++ src/lib/mappings.ts
@@ -1,0 +1,23 @@
+/**
+ * Temporary mapping for tech stack icons.
+ */
+const techStackMap: Record<string, string> = {
+    'React': 'reactjs',
+    'TypeScript': 'typescript',
+    'TailwindCSS': 'tailwindcss',
+    'Node.js': 'nodejs',
+    'NextJS': 'nextjs',
+    'aws': 'aws',
+    'MongoDB': 'mongodb',
+    'PostgreSQL': 'postgresql',
+    'Python': 'python',
+    'Flask': 'flask',
+    'Solidity': 'solidity',
+    'C#': 'csharp',
+    'Firebase': 'firebase',
+    'AWS IoT': 'aws',
+    'C++': 'c++',
+    'Azure': 'azure'
+}
+
+export default techStackMap;
